requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone-client
langchain-pinecone
langchain-community
langchain_openai
langchain_experimental 
